% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{analysis_complete}
\alias{analysis_complete}
\title{Is the analysis complete?}
\usage{
analysis_complete(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\value{
\code{TRUE} if the plate's analysis has been fully carried out;
\code{FALSE} otherwise.
}
\description{
Check if a ddPCR plate has been fully analyzed or if there are remaining steps.
}
\seealso{
\code{\link[ddpcr]{status}}\cr
\code{\link[ddpcr]{analyze}}
}

