% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{bind_df_ends}
\alias{bind_df_ends}
\title{Helper function for move_front and move_back}
\usage{
bind_df_ends(df, cols, dir = 1)
}
\description{
Helper function for move_front and move_back
}
\keyword{internal}

