% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-calculate_neg_freq.R
\name{calc_negative_freq_simple}
\alias{calc_negative_freq_simple}
\title{Calculate negative frequency based on number of drops}
\usage{
calc_negative_freq_simple(negative_drops, positive_drops)
}
\arguments{
\item{negative_drops}{Number of negative drops.}

\item{positive_drops}{Number of positive drops.}
}
\value{
Fraction of drops that are negative.
}
\description{
Calculate negative frequency based on number of drops
}
\examples{
calc_negative_freq_simple(5, 45)
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}}
}
\keyword{internal}

