% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentration.R
\name{calculate_concentration_single}
\alias{calculate_concentration_single}
\title{Calculate concentration in a single well}
\usage{
calculate_concentration_single(plate, well_id)
}
\value{
The concentration (integer) in a well.
}
\description{
Calculate concentration in a single well
}
\keyword{internal}
