% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{cat0}
\alias{cat0}
\title{Concatenate strings with no space between them}
\usage{
cat0(...)
}
\arguments{
\item{...}{Strings to concatenate}
}
\description{
Concatenate strings with no space between them
}
\keyword{internal}

