% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-classify.R
\name{classify_droplets.pnpp_experiment}
\alias{classify_droplets.pnpp_experiment}
\title{Analysis step: Classify droplets}
\usage{
\method{classify_droplets}{pnpp_experiment}(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\description{
Analysis step: Classify droplets
}
\keyword{internal}
