% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-classify.R
\name{classify_droplets_single.pnpp_experiment}
\alias{classify_droplets_single.pnpp_experiment}
\title{Classify droplets in a well}
\usage{
\method{classify_droplets_single}{pnpp_experiment}(plate, well_id, ...,
  plot = FALSE)
}
\description{
If you want to see details about how a well was classified, you can set
\code{plot = TRUE} to plot the results.
}
\keyword{internal}

