% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-classify.R
\name{classify_droplets_single}
\alias{classify_droplets_single}
\title{Classify droplets in a well}
\usage{
classify_droplets_single(plate, well_id, ...)
}
\description{
This function runs the actual algorithm for classifying droplets in a single
well.
}
\keyword{internal}

