% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{cluster}
\alias{cluster}
\title{Get cluster ID by name}
\usage{
cluster(plate, cluster)
}
\description{
Get cluster ID by name
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir)
# see what cluster names exist and their order
clusters(plate)
cluster(plate, 'FAILED')
cluster(plate, 'EMPTY')
}
}
\keyword{internal}

