% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{clusters}
\alias{clusters}
\title{Potential droplet clusters for a plate type}
\usage{
clusters(plate)
}
\arguments{
\item{plate}{a ddPCR plate.}
}
\value{
A character vector with the names of the clusters supported by the
plate type.
}
\description{
Each ddPCR plate type has a specific set of potential clusters the droplets
can be assigned to.
}
\details{
\href{https://github.com/daattali/ddpcr#extend}{See the README} for
more information on plate types.
}

