% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcr-package.R
\docType{package}
\name{ddpcr-package}
\alias{ddpcr}
\alias{ddpcr-package}
\title{ddpcr: Analysis and Visualization of Droplet Digital PCR in R and on the Web}
\description{
An interface to explore, analyze, and visualize droplet digital PCR (ddPCR) data in R. This is the first non-proprietary software for analyzing two-channel ddPCR data. An interactive tool was also created and is available online to facilitate this analysis for anyone who is not comfortable with using R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/daattali/ddpcr}
  \item \url{https://daattali.com/shiny/ddpcr/}
  \item Report bugs at \url{https://github.com/daattali/ddpcr/issues}
}

}
\author{
\strong{Maintainer}: Dean Attali \email{daattali@gmail.com} (\href{https://orcid.org/0000-0002-5645-3493}{ORCID})

}
\keyword{internal}
