% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ddpcr.R
\docType{package}
\name{ddpcr}
\alias{ddpcr}
\alias{ddpcr-package}
\title{ddPCR Package}
\description{
Analysis and visualization of Digital Droplet PCR data.
}
\details{
\href{http://daattali.com/shiny/ddpcr/}{Use the web tool} or
\href{https://github.com/daattali/ddpcr}{read the full documentation on GitHub}.
}

