% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-custom_thresholds.R
\name{define_clusters.custom_thresholds}
\alias{define_clusters.custom_thresholds}
\title{Define droplet clusters for custom thresholds plates}
\usage{
\method{define_clusters}{custom_thresholds}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define droplet clusters for custom thresholds plates
}
\keyword{internal}

