% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-ddpcr_plate.R
\name{define_clusters.ddpcr_plate}
\alias{define_clusters.ddpcr_plate}
\title{Define droplet clusters for default plates}
\usage{
\method{define_clusters}{ddpcr_plate}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define droplet clusters for default plates
}
\keyword{internal}
