% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{define_clusters.pnpp_experiment}
\alias{define_clusters.pnpp_experiment}
\title{Define droplet clusters for PNPP experiments}
\usage{
\method{define_clusters}{pnpp_experiment}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define droplet clusters for PNPP experiments
}

