% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-custom_thresholds.R
\name{define_params.custom_thresholds}
\alias{define_params.custom_thresholds}
\title{Define plate type parameters for custom thresholds plates}
\usage{
\method{define_params}{custom_thresholds}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define plate type parameters for custom thresholds plates
}
\keyword{internal}

