% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-ddpcr_plate.R
\name{define_params.ddpcr_plate}
\alias{define_params.ddpcr_plate}
\title{Define plate type parameters for default plates}
\usage{
\method{define_params}{ddpcr_plate}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define plate type parameters for default plates
}
\keyword{internal}

