% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-fam_positive_pnpp.R
\name{define_params.fam_positive_pnpp}
\alias{define_params.fam_positive_pnpp}
\title{Define plate type parameters}
\usage{
\method{define_params}{fam_positive_pnpp}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define plate type parameters
}

