% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-hex_positive_pnpp.R
\name{define_params.hex_positive_pnpp}
\alias{define_params.hex_positive_pnpp}
\title{Define plate type parameters for HEX-positive PNPP}
\usage{
\method{define_params}{hex_positive_pnpp}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define plate type parameters for HEX-positive PNPP
}
\keyword{internal}
