% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-kras.R
\name{define_params.kras}
\alias{define_params.kras}
\title{Define default parameters of KRAS assay}
\usage{
\method{define_params}{kras}(plate)
}
\description{
Define default parameters of KRAS assay
}
\seealso{
\code{\link[ddpcrS3]{define_params}},
\code{\link[ddpcrS3]{KRAS}}
}
\keyword{internal}

