% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{define_params.pnpp_experiment}
\alias{define_params.pnpp_experiment}
\title{Define plate type parameters for PNPP experiments}
\usage{
\method{define_params}{pnpp_experiment}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define plate type parameters for PNPP experiments
}
\keyword{internal}
