% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-wildtype_mutant_pnpp.R
\name{define_params.wildtype_mutant_pnpp}
\alias{define_params.wildtype_mutant_pnpp}
\title{Define plate type parameters for wildtype/mutant PNPP}
\usage{
\method{define_params}{wildtype_mutant_pnpp}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define plate type parameters for wildtype/mutant PNPP
}
\keyword{internal}

