% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-wtnegbraf.R
\name{define_params.wtnegbraf}
\alias{define_params.wtnegbraf}
\title{Define default parameters of WT-NEG-BRAF assay}
\usage{
\method{define_params}{wtnegbraf}(plate)
}
\description{
Define default parameters of WT-NEG-BRAF assay
}
\seealso{
\code{\link[ddpcrS3]{define_params}},
\code{\link[ddpcrS3]{WTNEGBRAF}}
}
\keyword{internal}

