% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-custom_thresholds.R
\name{define_steps.custom_thresholds}
\alias{define_steps.custom_thresholds}
\title{Define analysis steps for custom thresholds plates}
\usage{
\method{define_steps}{custom_thresholds}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define analysis steps for custom thresholds plates
}
\keyword{internal}

