% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-ddpcr_plate.R
\name{define_steps.ddpcr_plate}
\alias{define_steps.ddpcr_plate}
\title{Define analysis steps for default plates}
\usage{
\method{define_steps}{ddpcr_plate}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define analysis steps for default plates
}

