% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{define_steps.pnpp_experiment}
\alias{define_steps.pnpp_experiment}
\title{Define analysis steps for PNPP experiments}
\usage{
\method{define_steps}{pnpp_experiment}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Define analysis steps for PNPP experiments
}
\keyword{internal}

