% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{diff.point2d}
\alias{diff.point2d}
\title{Euclidean distance between two points (if second point is not given,
calculate distance to the origin)}
\usage{
\method{diff}{point2d}(v, w)
}
\description{
Euclidean distance between two points (if second point is not given,
calculate distance to the origin)
}
\keyword{internal}

