% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-fam_positive_ppnp.R
\docType{data}
\name{FAM_POSITIVE_PPNP}
\alias{FAM_POSITIVE_PPNP}
\title{WT-NEG-BRAF assay}
\format{\preformatted{ chr "fam_positive_ppnp"
}}
\usage{
FAM_POSITIVE_PPNP
}
\description{
"Wild-type negative BRAF" assay type. Use this when initializing a ddPCR
plate that uses the WT-NEG-BRAF assay.
}
\details{
TODO what this assay looks like
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcr")
new_plate(dir = dir, type = FAM_POSITIVE_PPNP)
}
\seealso{
\code{\link[ddpcr]{new_plate}},
\code{\link[ddpcr]{PPNP_ASSAy}}
}
\keyword{datasets}

