% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{format.point2d}
\alias{format.point2d}
\title{Format a point2d for pretty printing}
\usage{
\method{format}{point2d}(x, ...)
}
\description{
Format a point2d for pretty printing
}
\keyword{internal}

