% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pnpp_experiment-empty.R
\name{get_empty_cutoff.pnpp_experiment}
\alias{get_empty_cutoff.pnpp_experiment}
\title{Get the cutoff for empty droplets in a well}
\usage{
\method{get_empty_cutoff}{pnpp_experiment}(plate, well_id)
}
\description{
Very similar to the \code{get_empty_cutoff} method of the base plate type,
except for this plate type we know that we don't expect any droplets in one
of the corners, so we can save time by only having an empty cutoff in one
dimension.
}
\keyword{internal}

