% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/empty.R
\name{get_empty_cutoff}
\alias{get_empty_cutoff}
\title{Get the cutoff for empty droplets in a well}
\usage{
get_empty_cutoff(plate, well_id)
}
\value{
A list with 2 elements named \code{x} and \code{y} with the values
being the cutoff in the corresponding axis.
}
\description{
Calculate the cutoff thresholds that define which droplets are empty in a well.
}
\keyword{internal}

