% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ppnp_assay-filled.R
\name{get_filled_borders}
\alias{get_filled_borders}
\title{Get borders of filled droplets in PPNP assay}
\usage{
get_filled_borders(plate, well_id)
}
\arguments{
\item{plate}{A ddPCR plate.}

\item{well_id}{Get borders of filled droplets for this well.}
}
\value{
A 2-element numeric vector defining the thresholds of filled drops
in the positive dimension.
}
\description{
In a PPNP assay, the rain droplets are the non-empty drops that don't have a
high enough intensity in the positive dimension to be considered as filled
with high quality sample DNA. Only droplets considered as filled are
candidates for the \code{negative} and \code{positive} clusters.
\code{get_filled_borders} returns the threshold values in the positive
dimension that are used to determine which drops are filled.
}
\examples{
file <- system.file("sample_data", "small", "analyzed_ppnp.rds", package = "ddpcrS3")
plate <- load_plate(file)
get_filled_borders(plate, "B06")
get_filled_borders(plate, "C09")
}
\seealso{
\code{\link[ddpcrS3]{PPNP_ASSAy}},

\code{\link[ddpcrS3]{positive_dim}},

\code{\link[ddpcrS3]{get_filled_drops}}
}
\keyword{internal}

