% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ppnp_assay-filled.R
\name{get_filled_drops}
\alias{get_filled_drops}
\title{Get filled droplets in PPNP assay}
\usage{
get_filled_drops(plate, well_id, borders)
}
\arguments{
\item{plate}{A ddPCR plate.}

\item{well_id}{Get borders of filled droplets for this well.}

\item{borders}{(Optional) The filled droplets borders, as calculated by
\code{\link[ddpcrS3]{get_filled_drops}}. If missing, then
\code{\link[ddpcrS3]{get_filled_drops}} is called to calculate the borders.}
}
\value{
Dataframe with all filled droplets in the given well.
}
\description{
In a PPNP assay, the rain droplets are the non-empty drops that don't have a
high enough intensity in the positive dimension to be considered as filled
with high quality sample DNA. Only droplets considered as filled are
candidates for the \code{negative} and \code{positive} clusters.
\code{get_filled_drops} returns the droplets that are considered filled.
}
\examples{
file <- system.file("sample_data", "small", "analyzed_ppnp.rds", package = "ddpcrS3")
plate <- load_plate(file)
get_filled_drops(plate, "B06")
get_filled_drops(plate, "B06", get_filled_borders(plate, "B06"))
}
\seealso{
\code{\link[ddpcrS3]{PPNP_ASSAy}},

\code{\link[ddpcrS3]{positive_dim}},

\code{\link[ddpcrS3]{get_filled_drops}}
}
\keyword{internal}

