% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inflection_pts}
\alias{get_inflection_pts}
\title{Get the indices of the inflection points in a curve}
\usage{
get_inflection_pts(dat)
}
\arguments{
\item{dat}{Dataframe containing an \code{x} and \code{y} variable.}
}
\description{
Get the indices of the inflection points in a curve
}
\examples{
curve1 <- dplyr::data_frame(x = -10:3, y = (x + 5) ^ 3 - 10)
curve1[get_inflection_pts(curve1), ]$x
}
\keyword{internal}

