% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outliers.R
\name{get_outlier_cutoff.ddpcr_plate}
\alias{get_outlier_cutoff.ddpcr_plate}
\title{Get the cutoff for outliers}
\usage{
\method{get_outlier_cutoff}{ddpcr_plate}(plate)
}
\description{
Get the cutoff for outliers
}
\keyword{internal}

