% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers.R
\name{get_outlier_cutoff}
\alias{get_outlier_cutoff}
\title{Get the cutoff for outliers}
\usage{
get_outlier_cutoff(plate)
}
\value{
A named list with two elements, giving the cutoff for outliers in
each dimension.
}
\description{
Get the cutoff for outliers
}
\keyword{internal}

