% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_single_well}
\alias{get_single_well}
\title{Get droplet data from a well}
\usage{
get_single_well(plate, well_id, empty = FALSE, outliers = FALSE,
  clusters = FALSE)
}
\arguments{
\item{plate}{}
}
\description{
Get droplet data from a well
}
\keyword{internal}

