% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{get_wells_btwn}
\alias{get_wells_btwn}
\title{Get all wells between two wells (assume a rectangle layout)}
\usage{
get_wells_btwn(well1, well2)
}
\description{
Get all wells between two wells (assume a rectangle layout)
}
\examples{
get_wells_btwn("C04", "D06")
}
\keyword{internal}
