% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%btwn\%}
\alias{\%btwn\%}
\title{Determine if a numeric value is within a range}
\usage{
x \%btwn\% rng
}
\arguments{
\item{x}{Numeric vector to check whether the values are within \code{rng}}

\item{rng}{The range to check if numbers in \code{x} are within it.}
}
\description{
Determine if a numeric value is within a range
}
\keyword{internal}
