% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shine.R
\name{gui}
\alias{gui}
\title{Run the interactive analysis tool (Shiny app) in a web browser}
\usage{
gui()
}
\description{
In addition to the functions provided in this package, the \code{ddpcr} package
also provides an interactive tool that can be used to analyze ddPCR data
more easily. The tool will be launched in a web browser.
}

