% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{has_step}
\alias{has_step}
\title{Does a ddPCR plate have a step with this name?}
\usage{
has_step(plate, step)
}
\arguments{
\item{plate}{A ddPCR plate}

\item{step}{A step name}
}
\description{
Does a ddPCR plate have a step with this name?
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir)
steps(plate)
has_step(plate, 'REMOVE_FAILURES')
has_step(plate, 'NO_SUCH_STEP')
}
}
\keyword{internal}

