% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex-plot.R
\name{hex_plot.ddpcr_plate}
\alias{hex_plot.ddpcr_plate}
\title{Quickly plot a ddPCR plate with hex-based binning}
\usage{
\method{hex_plot}{ddpcr_plate}(x, wells, samples, superimpose = FALSE,
  show_full_plate = FALSE, show_failed_wells = TRUE,
  alpha_bg_failed = 0.7, bg_plot = "transparent", bg_failed = "#111111",
  bg_unused = "#FFFFFF", xlab = x_var(plate), ylab = y_var(plate),
  title = NULL, show_grid = FALSE, show_grid_labels = FALSE,
  text_size_title = 14, text_size_row_col = 12,
  text_size_axes_labels = 12, text_size_grid_labels = 12, ...)
}
\arguments{
\item{x}{A ddPCR plate.}

\item{wells}{Only plot selected wells. Supports range notation, see
\code{\link[ddpcr]{subset.ddpcr_plate}}.}

\item{samples}{Only plot selected samples.}

\item{superimpose}{If \code{TRUE}, show all wells superimposed in one plot;
otherwise, show wells in a grid.}

\item{show_full_plate}{If \code{TRUE}, show full 96-well plate; otherwise,
show only plate rows and columns that are used.}

\item{show_failed_wells}{Whether or not to include wells that are deemed
as failed ddPCR runs.}

\item{alpha_bg_failed}{The transparency of the background of failed wells.}

\item{bg_plot}{The background colour for the plot.}

\item{bg_failed}{The background colour to use for failed wells.}

\item{bg_unused}{The background colour to use for unused wells.}

\item{xlab}{The label on the X axis.}

\item{ylab}{The label on the Y axis.}

\item{title}{The title for the plot.}

\item{show_grid}{Whether or not to show grid lines.}

\item{show_grid_labels}{Whether or not to show numeric labels for the grid
lines along the axes.}

\item{text_size_title}{Text size of the title.}

\item{text_size_row_col}{Text size of the row and column labels.}

\item{text_size_axes_labels}{Text size of the X/Y axis labels.}

\item{text_size_grid_labels}{Text size of the numeric grid line labels.}

\item{...}{Ignored.}
}
\value{
A ggplot2 plot object.
}
\description{
Plot the data of a ddPCR plate. A plate can be plotted throughout any stage
of the analysis, and the most up-to-date data will be shown. For example,
a plot performed after initializing a plat will show all the raw data, but
a plot performed after analyzing a plate will show information such as failed
wells.
}
\section{Extending ddpcr_plate}{

If you create your own plate type, this default hex plot function might be
enough if there is no extra information you want to display in a plot.
If you do need to provide a more customized plot function, it can be
a good idea to use the output from this plot function as a basis and only
add the code that is necessary to append to the plot.
}
\examples{
\dontrun{
plate <- new_plate(sample_data_dir())
hex_plot(plate)
plate <- plate \%>\% analyze
hex_plot(plate)
}
}

