% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-hex_positive_ppnp.R
\docType{data}
\name{HEX_POSITIVE_PPNP}
\alias{HEX_POSITIVE_PPNP}
\title{KRAS assay}
\format{\preformatted{ chr "hex_positive_ppnp"
}}
\usage{
HEX_POSITIVE_PPNP
}
\description{
"KRAS" assay type. Use this when initializing a ddPCR
plate that uses the KRAS assay.
}
\details{
TODO what this assay looks like
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcr")
new_plate(dir = dir, type = HEX_POSITIVE_PPNP)
}
\seealso{
\code{\link[ddpcr]{new_plate}},
\code{\link[ddpcr]{PPNP_ASSAy}}
}
\keyword{datasets}

