% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_dir}
\alias{is_dir}
\title{Determine if a given path is a valid directory}
\usage{
is_dir(path)
}
\arguments{
\item{path}{A file path to test}
}
\description{
Determine if a given path is a valid directory
}
\keyword{internal}
