% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{is_dirty}
\alias{is_dirty}
\title{Is the plate object dirty (ie has changed since the analysis was run)?}
\usage{
is_dirty(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\value{
\code{TRUE} if any plate settings have changed that require the plate
analysis to re-run; \code{FALSE} otherwise
}
\description{
Is the plate object dirty (ie has changed since the analysis was run)?
}
\keyword{internal}

