% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{is_empty_plate}
\alias{is_empty_plate}
\title{Is a plate empty?}
\usage{
is_empty_plate(plate)
}
\description{
A plate is considered empty if it has not yet been fully initialized, and thus
its status is 0.
}
\seealso{
\code{\link[ddpcr]{status}}
}
\keyword{internal}

