% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_file}
\alias{is_file}
\title{Determine if a given path is a valid file}
\usage{
is_file(path)
}
\arguments{
\item{path}{A file path to test}
}
\description{
Determine if a given path is a valid file
}
\keyword{internal}
