% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{is_range}
\alias{is_range}
\title{Is the given parameter a range?}
\usage{
is_range(x)
}
\description{
Is the given parameter a range?
}
\examples{
is_range("C05")            # FALSE
is_range(c("C05", "C08"))  # FALSE
is_range("C05")            # FALSE
is_range("C05, C08")       # TRUE
is_range("C05:C08")        # TRUE
is_range("C05.C08")        # FALSE
}
\keyword{internal}

