% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/failures.R
\name{is_well_success.ddpcr_plate}
\alias{is_well_success.ddpcr_plate}
\title{Algorithm for determining if a single well failed}
\usage{
\method{is_well_success}{ddpcr_plate}(plate, well_id)
}
\description{
Algorithm for determining if a single well failed
}
\keyword{internal}

