% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-failures.R
\name{is_well_success.pnpp_experiment}
\alias{is_well_success.pnpp_experiment}
\title{Determine if a well had a successful ddPCR run}
\usage{
\method{is_well_success}{pnpp_experiment}(plate, well_id)
}
\description{
Very similar to the \code{is_well_success} method of the base plate type,
except for this plate type we know where we expect to see droplets so we 
can save time by using QC metrics more specific to this plate type.
}
\keyword{internal}

