% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/failures.R
\name{is_well_success}
\alias{is_well_success}
\title{Determine is a well was a success or failure}
\usage{
is_well_success(plate, well_id)
}
\description{
Determine is a well was a success or failure
}
\keyword{internal}

