% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-kras.R
\docType{data}
\name{KRAS}
\alias{KRAS}
\title{KRAS assay}
\format{\preformatted{ chr "kras"
}}
\usage{
KRAS
}
\description{
"KRAS" assay type. Use this when initializing a ddPCR
plate that uses the KRAS assay.
}
\details{
TODO what this assay looks like
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcrS3")
new_plate(dir = dir, type = KRAS)
new_plate(dir = dir, type = "kras")
}
\seealso{
\code{\link[ddpcrS3]{new_plate}},
\code{\link[ddpcrS3]{PPNP_ASSAy}}
}
\keyword{datasets}

