% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{local_minima}
\alias{local_minima}
\title{Get the indices of the local minima in a list of numbers}
\usage{
local_minima(x)
}
\description{
Get the indices of the local minima in a list of numbers
}
\keyword{internal}

