% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{lol_to_df}
\alias{lol_to_df}
\title{Convert a list of lists outputted from vapply to dataframe}
\usage{
lol_to_df(lol, name = "well")
}
\arguments{
\item{lol}{List of lists that is a result of a vapply}

\item{name}{Column name to use for the name of each list}
}
\description{
Convert a list of lists outputted from vapply to dataframe
}
\examples{
vapply(c("a", "b", "c"),
       function(x) list(low = x, up = toupper(x)),
       list(character(1), character(1))) \%>\%
  lol_to_df("key")
}
\keyword{internal}

