% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-classify.R
\name{mark_clusters}
\alias{mark_clusters}
\title{Mark the clusters of droplets only in certain wells to their assigned cluster}
\usage{
mark_clusters(plate, wells)
}
\description{
This function simply looks at all droplets of certain wells and marks the
cluster of each droplet according to the gates that are already calculated.
This function is called once after \code{\link[ddpcr]{classify_droplets_single}}
determines the gates for every well, and it's called again when reclassifying
wells gives us more accurate information.
}
\keyword{internal}

