% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{move_back}
\alias{move_back}
\title{Move columns to the back of a data.frame (taken from daattali/rsalad)}
\usage{
move_back(df, cols)
}
\description{
Move columns to the back of a data.frame (taken from daattali/rsalad)
}
\keyword{internal}

