% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_front}
\alias{move_front}
\title{Move columns to the front of a data.frame}
\usage{
move_front(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{A vector of column names to move to the front.}
}
\description{
This function is taken from daattali/rsalad R package.
}
\examples{
df <- data.frame(a = character(0), b = character(0), c = character(0))
move_front(df, "b")
move_front(df, c("c", "b"))
}
\keyword{internal}

