% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msg}
\alias{msg}
\title{Write a message to the user if the `ddpcr.verbose` option is on}
\usage{
msg(...)
}
\arguments{
\item{...}{Parameters to pass to \code{message()}}
}
\description{
Running a ddpcr analysis results in many messages being printed to the console.
By default, these messages are on when the user is using R interactively
and off otherwise. You can overwrite this setting with \code{options(ddpcr.verbose = FALSE)}.
}
\keyword{internal}

