% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/save_load.R
\name{normalize_to_rds}
\alias{normalize_to_rds}
\title{Normalize a file name to .rds suffix}
\usage{
normalize_to_rds(file)
}
\description{
Normalize a file name to .rds suffix
}
\examples{
normalize_to_rds("somefile")       # somefile.rds
normalize_to_rds("somefile.rds")   # somefile.rds
normalize_to_rds("somefile.r")     # somefile.r.rds
}
\keyword{internal}

