% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\name{other_dim}
\alias{other_dim}
\title{Given an axis (X or Y), return the other}
\usage{
other_dim(dim)
}
\description{
Given an axis (X or Y), return the other
}
\examples{
other_dim("X")
other_dim("y")
}
\keyword{internal}

