% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-ddpcr_plate.R
\name{parent_plate_type.ddpcr_plate}
\alias{parent_plate_type.ddpcr_plate}
\title{Parent plate type of default plates}
\usage{
\method{parent_plate_type}{ddpcr_plate}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Parent plate type of default plates
}

