% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-ddpcr_plate.R
\name{parent_plate_type.default}
\alias{parent_plate_type.default}
\title{Parent plate type of any plate}
\usage{
\method{parent_plate_type}{default}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Parent plate type of any plate
}
\keyword{internal}
