% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-fam_positive_pnpp.R
\name{parent_plate_type.fam_positive_pnpp}
\alias{parent_plate_type.fam_positive_pnpp}
\title{Parent plate type}
\usage{
\method{parent_plate_type}{fam_positive_pnpp}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Parent plate type
}

