% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-kras.R
\name{parent_plate_type.kras}
\alias{parent_plate_type.kras}
\title{Define parent plate type of KRAS assay}
\usage{
\method{parent_plate_type}{kras}(plate)
}
\description{
Define parent plate type of KRAS assay
}
\seealso{
\code{\link[ddpcrS3]{parent_plate_type}},
\code{\link[ddpcrS3]{KRAS}}
}
\keyword{internal}

