% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-wildtype_mutant_pnpp.R
\name{parent_plate_type.wildtype_mutant_pnpp}
\alias{parent_plate_type.wildtype_mutant_pnpp}
\title{Parent plate type of wildtype/mutant PNPP}
\usage{
\method{parent_plate_type}{wildtype_mutant_pnpp}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Parent plate type of wildtype/mutant PNPP
}
\keyword{internal}

