% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-wtnegbraf.R
\name{parent_plate_type.wtnegbraf}
\alias{parent_plate_type.wtnegbraf}
\title{Define parent plate type of WT-NEG-BRAF assay}
\usage{
\method{parent_plate_type}{wtnegbraf}(plate)
}
\description{
Define parent plate type of WT-NEG-BRAF assay
}
\seealso{
\code{\link[ddpcrS3]{parent_plate_type}},
\code{\link[ddpcrS3]{WTNEGBRAF}}
}
\keyword{internal}

