% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{plate_data<-}
\alias{plate_data<-}
\title{Overwrite the plate data}
\usage{
plate_data(plate) <- value
}
\arguments{
\item{plate}{A ddPCR plate}

\item{value}{New plate data}
}
\description{
When creating custom analysis steps for new plate types, it is often necessary
to set the plate data with new data, especially when assigning new clusters
to the plate droplets.
}
\seealso{
\code{\link[ddpcr]{plate_data}}
}
\keyword{internal}
