% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{plate_meta<-}
\alias{plate_meta<-}
\title{Overwrite the plate metadata}
\usage{
plate_meta(plate) <- value
}
\arguments{
\item{plate}{A ddPCR plate}

\item{value}{New plate metadata}
}
\description{
When creating custom analysis steps for new plate types, it is often necessary
to add/change variables in the plate metadata.
}
\seealso{
\code{\link[ddpcr]{plate_meta}}
}
\keyword{internal}

