% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{point2d}
\alias{point2d}
\title{Representation of a 2D point}
\usage{
point2d(x)
}
\arguments{
\item{x}{A 2-element numeric vector.}
}
\value{
An object of class \code{point2d} with the given coordinates.
}
\description{
Representation of a 2D point
}
\examples{
point2d(c(10, 20))
}
\keyword{internal}

