% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\name{positive_dim}
\alias{positive_dim}
\alias{positive_dim<-}
\title{Positive dimension in a PPNP assay}
\usage{
positive_dim(plate)

positive_dim(plate) <- value
}
\description{
Get or set the dimension (X or Y) that has a high intensity in all non-empty
drops in a \code{PPNP_ASSAY}.
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcrS3")
plate <- new_plate(dir = dir, type = PPNP_ASSAY)
positive_dim(plate) <- "Y"
}
\seealso{
\code{\link[ddpcrS3]{PPNP_ASSAy}},
\code{\link[ddpcrS3]{variable_dim}}
}

