% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{positive_name}
\alias{positive_name}
\alias{positive_name<-}
\alias{negative_name}
\alias{negative_name<-}
\title{Name identifier for positive and negative droplets}
\usage{
positive_name(plate)

positive_name(plate) <- value

negative_name(plate)

negative_name(plate) <- value
}
\arguments{
\item{plate}{A ddPCR plate.}

\item{value}{The identifier to use for droplets in the positive/negative cluster.}
}
\description{
Name identifier for positive and negative droplets
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}}
}
\keyword{internal}
