% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\docType{data}
\name{PPNP_ASSAY}
\alias{PPNP_ASSAY}
\title{PPNP assay}
\format{\preformatted{ chr "ppnp_assay"
}}
\usage{
PPNP_ASSAY
}
\description{
"Positive-Positive;Negative-Positive assay" type. Use this assay when
your ddPCR data has 3 main clusters: one cluster near the bottom left that
corresponds to empty drops, one cluster near the top right with drops that
have a high intensity in both x and y dimensions (Positive-Positive), and one
cluster at one of the other two corners with drops that have a high intensity
in one diemnsion but a low intensity in the other (Negative-Positive).
}
\details{
TODO what this assay looks like (POSITIVE_DIMENSION must be set, either
in params or by defining a child type)
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcrS3")
new_plate(dir = dir, type = PPNP_ASSAY)
new_plate(dir = dir, type = "wtnegbraf")
}
\seealso{
\code{\link[ddpcrS3]{new_plate}},
\code{\link[ddpcrS3]{positive_dim}},
\code{\link[ddpcrS3]{wells_positive}},
\code{\link[ddpcrS3]{wells_negative}}
}
\keyword{datasets}

