% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{quiet}
\alias{quiet}
\title{Suppress all output from an expression. Works cross-platform.}
\usage{
quiet(expr, all = TRUE)
}
\arguments{
\item{expr}{Expression to run.}

\item{all}{If \code{TRUE} then suppress warnings and messages as well;
otherwise, only suppress printed output (such as from \code{print} or
\code{cat}).}
}
\description{
Suppress all output from an expression. Works cross-platform.
}
\keyword{internal}

