% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subset.R
\name{range_list_to_vec}
\alias{range_list_to_vec}
\title{Convert a list of ranges to a vector of its individual components}
\usage{
range_list_to_vec(rangel)
}
\description{
Convert a list of ranges to a vector of its individual components
}
\examples{
range_list_to_vec("A01")
range_list_to_vec("A01:A04")
range_list_to_vec("A01, B03")
range_list_to_vec("A01, B02:C04, C07")
}
\keyword{internal}

