% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{range_to_endpoints}
\alias{range_to_endpoints}
\title{Extract the two endpoints of a range}
\usage{
range_to_endpoints(range)
}
\description{
Extract the two endpoints of a range
}
\examples{
range_to_endpoints("B05:G09")   # c("B05", "G09")
range_to_endpoints("B05")       # c("B05", "B05")
}
\keyword{internal}
