% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subset.R
\name{range_to_seq}
\alias{range_to_seq}
\title{Convert a range to a vector of all elements between the endpoints}
\usage{
range_to_seq(rng)
}
\description{
Convert a range to a vector of all elements between the endpoints
}
\examples{
range_to_seq(c(5, 8))   # 5:8
range_to_seq(c(8, 5))   # 5:8
}
\keyword{internal}

