% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pnpp_experiment-reclassify.R
\name{reclassify_droplets.pnpp_experiment}
\alias{reclassify_droplets.pnpp_experiment}
\title{Analysis step: Reclassify droplets}
\usage{
\method{reclassify_droplets}{pnpp_experiment}(plate)
}
\description{
Analysis step: Reclassify droplets
}
\keyword{internal}

