% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-reclassify.R
\name{reclassify_droplets_single.pnpp_experiment}
\alias{reclassify_droplets_single.pnpp_experiment}
\title{Reclassify droplets in a well}
\usage{
\method{reclassify_droplets_single}{pnpp_experiment}(plate, well_id, ...,
  consensus_border_ratio)
}
\description{
Reclassify droplets in a well given the ratio of where to place the MT border
over the median WT drops
}
\keyword{internal}
