% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/empty.R
\name{remove_empty.ddpcr_plate}
\alias{remove_empty.ddpcr_plate}
\title{Remove empty droplets}
\usage{
\method{remove_empty}{ddpcr_plate}(plate)
}
\description{
The algorithm for removing empty droplets from a plate
}
\keyword{internal}

