% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pnpp_experiment-empty.R
\name{remove_empty.pnpp_experiment}
\alias{remove_empty.pnpp_experiment}
\title{Remove empty droplets}
\usage{
\method{remove_empty}{pnpp_experiment}(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\description{
Remove empty droplets
}
\keyword{internal}

