% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{remove_failures.ddpcr_plate}
\alias{remove_failures.ddpcr_plate}
\title{Analysis step: Remove failed wells}
\usage{
\method{remove_failures}{ddpcr_plate}(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\description{
Analysis step: Remove failed wells
}
\keyword{internal}

