% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_data_dir}
\alias{sample_data_dir}
\title{Get directory of sample ddPCR data}
\usage{
sample_data_dir()
}
\value{
directory of sample ddPCR data
}
\description{
This function returns a directory that contains sample ddPCR data, which can
be used to initialize a ddPCR plate. It is mainly used to have an easy way
to initialize ddpcr plates in the documentation examples, but you can also
use it for learning purposes.
}
\examples{
new_plate(dir = sample_data_dir())
}

