% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plate.R
\name{set_default_params}
\alias{set_default_params}
\title{Reset plate parameters to their defaults}
\usage{
set_default_params(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\value{
The plate with the parameters set to the plate type's default values.
}
\description{
Use this function to reset a ddPCR plate's parameters back to their default
values.
}
\seealso{
\code{\link[ddpcr]{params}}
}

