% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{status}
\alias{status}
\alias{status<-}
\title{Plate status}
\usage{
status(plate)

status(plate) <- value
}
\description{
The status of a plate corresponds to the number of analysis steps that have
taken place. A plate that has been initialized but has not yet been analyzed
at all has status 1.\cr\cr
If you add custom analysis steps to a new plate type, you should make sure
to update the status of the plate after each step. You can use 
\code{\link[ddpcr]{check_step}} to ensure that the plate is at an appropriate
status before beginning each step.
}
\seealso{
\code{\link[ddpcr]{steps}}\cr
\code{\link[ddpcr]{check_step}}
}
\keyword{internal}

