% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{step}
\alias{step}
\title{Get step ID by step name}
\usage{
step(plate, step)
}
\description{
Get step ID by step name
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir)
# see what step names exist and their order
steps(plate)
step(plate, 'REMOVE_OUTLIERS')
}
}
\keyword{internal}

