% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-extend.R
\name{step_begin}
\alias{step_begin}
\title{Inform the user that an analysis step is starting}
\usage{
step_begin(text)
}
\arguments{
\item{text}{The text to show the user when this step begins.}
}
\description{
When an analysis step starts running, it's recommended to call this function
so that the user will know what step is taking place. The time when a step
begins gets recorded so that the user will see exactly how long it took.
}
\seealso{
\code{\link[ddpcr]{step_end}}
}
\keyword{internal}

