% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate-extend.R
\name{step_end}
\alias{step_end}
\title{Inform the user that an analysis step finished}
\usage{
step_end()
}
\description{
When an analysis step is done, it's recommended to call this function
so that the user will know the step finished. The time when a step
finishes gets recorded so that the user will see exactly how long it took.
An earlier call to \code{step_begin} must have taken place.
}
\seealso{
\code{\link[ddpcr]{step_begin}}
}
\keyword{internal}
