% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{step_name}
\alias{step_name}
\title{Get step name by ID}
\usage{
step_name(plate, step)
}
\description{
Get step name by ID
}
\examples{
\dontrun{
plate <- new_plate(sample_data_dir())
# see what step names exist and their order
steps(plate)
step_name(plate, 2)
}
}
\keyword{internal}

