% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{variable_dim}
\alias{variable_dim}
\alias{variable_dim<-}
\title{Variable dimension in a PNPP experiment}
\usage{
variable_dim(plate)

variable_dim(plate) <- value
}
\description{
Get or set the variable dimension (X or Y), which is defined as the dimension
that can have both high and low fluorescence intensities in the non-empty
drops in a \code{pnpp_experiment} plate.
}
\examples{
plate <- new_plate(dir = sample_data_dir(), type = plate_types$pnpp_experiment)
variable_dim(plate) <- "Y"
variable_dim(plate)
positive_dim(plate)
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}}\cr
\code{\link[ddpcr]{positive_dim}}
}
\keyword{internal}
