% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/subset.R
\docType{data}
\name{WELL_ID_REGEX}
\alias{WELL_ID_REGEX}
\title{regex for a well ID}
\format{\preformatted{ chr "^[A-H][0-1][0-9]$"
}}
\usage{
WELL_ID_REGEX
}
\description{
regex for a well ID
}
\keyword{internal}

