% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{wells_negative}
\alias{wells_negative}
\title{Get wells containing many negative droplets in a PNPP experiment}
\usage{
wells_negative(x)
}
\arguments{
\item{x}{A ddPCR plate of type pnpp_experiment}
}
\value{
Vector of wells with significant number of negative drops.
}
\description{
In a PNPP experiment, the two main non-empty clusters of drops are the negative
and positive clusters. The positive cluster appears in every well, while the
negative cluster is not necessarily in every well. \code{wells_negative}
returns a list of wells that have a significant number of negative drops.
}
\examples{
file <- system.file("sample_data", "small", "analyzed_pnpp.rds", package = "ddpcr")
plate <- load_plate(file)
plate \%>\% wells_negative
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}},
\code{\link[ddpcr]{wells_positive}}
}

