% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-wtnegbraf.R
\docType{data}
\name{WTNEGBRAF}
\alias{WTNEGBRAF}
\title{WT-NEG-BRAF assay}
\format{\preformatted{ chr "wtnegbraf"
}}
\usage{
WTNEGBRAF
}
\description{
"Wild-type negative BRAF" assay type. Use this when initializing a ddPCR
plate that uses the WT-NEG-BRAF assay.
}
\details{
TODO what this assay looks like
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcr")
new_plate(dir = dir, type = WTNEGBRAF)
new_plate(dir = dir, type = "wtnegbraf")
}
\seealso{
\code{\link[ddpcr]{new_plate}},
\code{\link[ddpcr]{PPNP_ASSAy}}
}
\keyword{datasets}

