% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-calculate_neg_freq.R
\name{calculate_neg_freq_single}
\alias{calculate_neg_freq_single}
\title{Calculate negative frequency of a single well}
\usage{
calculate_neg_freq_single(plate, well_id)
}
\arguments{
\item{plate}{A ddPCR plate}

\item{well_id}{A well ID}
}
\value{
list with 3 elemnts: number of negative drops, number of positive
drops, and fraction of negative drops.
}
\description{
Calculate negative frequency of a single well
}
\examples{
file <- system.file("sample_data", "small", "analyzed_pnpp.rds", package = "ddpcr")
plate <- load_plate(file)
plate \%>\% calculate_neg_freq_single("B06")
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}}
}
\keyword{internal}

