% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddpcr.R
\docType{package}
\name{ddpcr}
\alias{ddpcr}
\alias{ddpcr-package}
\title{Analysis and visualization of Droplet Digital PCR data.}
\description{
An interface to explore, analyze, and visualize droplet digital PCR
(ddPCR) data in R. This is the first known non-proprietary software for
analyzing ddPCR data. An interactive tool was also created and is available
online to facilitate this analysis for anyone who is not comfortable with
using R.\cr\cr
\href{http://daattali.com/shiny/ddpcr/}{Use the web tool} or
\href{https://github.com/daattali/ddpcr}{read the full documentation on GitHub}.
}
