% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{diff.point2d}
\alias{diff.point2d}
\title{Euclidean distance between two points}
\usage{
\method{diff}{point2d}(x, y, ...)
}
\arguments{
\item{x,y}{Points generated with \code{point2d}}
}
\description{
Calculate the distance between two points in 2D space. If only one points is
given, then the distance to the origin is calculated.
}
\seealso{
\code{\link[ddpcrS3]{point2d}}
}
\keyword{internal}

