% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{get_empty_cutoff.ddpcr_plate}
\alias{get_empty_cutoff.ddpcr_plate}
\title{Get the cutoff for empty droplets in a well}
\usage{
\method{get_empty_cutoff}{ddpcr_plate}(plate, well_id)
}
\description{
Get the cutoff for empty droplets in a well
}
\keyword{internal}

