% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/empty.R
\name{get_empty_cutoff}
\alias{get_empty_cutoff}
\title{Get the cutoff for empty droplets in a well}
\usage{
get_empty_cutoff(plate, well_id)
}
\description{
Get the cutoff for empty droplets in a well
}
\keyword{internal}

