% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inflection_pts}
\alias{get_inflection_pts}
\title{Get the indices of the inflection points in a curve}
\usage{
get_inflection_pts(dat)
}
\description{
Get the indices of the inflection points in a curve
}
\keyword{internal}

