% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outliers.R
\name{get_outlier_cutoff.ddpcr_plate}
\alias{get_outlier_cutoff.ddpcr_plate}
\title{Algorithm for determining the cutoff for outliers}
\usage{
\method{get_outlier_cutoff}{ddpcr_plate}(plate)
}
\description{
Algorithm for determining the cutoff for outliers
}
\keyword{internal}

