% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-hex_positive_pnpp.R
\name{hex_positive_pnpp}
\alias{hex_positive_pnpp}
\title{Plate type: HEX-positive PNPP}
\usage{
plate_types$hex_positive_pnpp
}
\description{
A ddPCR plate of type \code{hex_positive_pnpp}, which can also be expressed as
(HEX+)/(FAM+HEX+), is a subtype of both \code{\link[ddpcr]{pnpp_experiment}}
and \code{\link[ddpcr]{wildtype_mutant_pnpp}}. Use this plate type if your data
has three main clusters of droplets: double-negative (empty droplets),
FAM+HEX+ (wildtype droplets) and HEX+FAM- (mutant droplets).
}
\details{
\href{https://github.com/daattali/ddpcr#extend}{See the README} for
more information on plate types.
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir, type = plate_types$hex_positive_pnpp)
type(plate)
plate \%>\% analyze \%>\% plot
}
}
\seealso{
\code{\link[ddpcr]{plate_types}}
\code{\link[ddpcr]{wildtype_mutant_pnpp}}
\code{\link[ddpcr]{fam_positive_pnpp}}
}

