% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/failures.R
\name{is_well_success.ddpcr_plate}
\alias{is_well_success.ddpcr_plate}
\title{Determine if a well had a successful ddPCR run}
\usage{
\method{is_well_success}{ddpcr_plate}(plate, well_id)
}
\description{
Determine if a well had a successful ddPCR run
}
\keyword{internal}

