% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{local_maxima}
\alias{local_maxima}
\title{Get the indices of the local maxima in a list of numbers}
\usage{
local_maxima(x)
}
\arguments{
\item{x}{Vector of numbers.}
}
\description{
Get the indices of the local maxima in a list of numbers
}
\examples{
local_maxima(c(1, 5, 3, 2, 4, 3))
}
\keyword{internal}

