% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{lol_to_df}
\alias{lol_to_df}
\title{Convert a list of lists outputted from vapply to dataframe}
\usage{
lol_to_df(lol, name = "well")
}
\description{
Convert a list of lists outputted from vapply to dataframe
}
\keyword{internal}

