% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\name{meta_var_name}
\alias{meta_var_name}
\title{Name of variable in PPNP assay metadata}
\usage{
meta_var_name(plate, var)
}
\description{
A default PPNP assay uses the names "positive" and "negative" for its two
non-empty clusters. If they are changed, then any variable in the metadata
will be renamed to use these names. \code{meta_var_name} translated a
default metadata variable name to the correct one.
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcrS3")
plate <- new_plate(dir = dir, type = PPNP_ASSAY)
params(plate, 'GENERAL', 'NEGATIVE_NAME') <- "mutant"
meta_var_name(plate, 'num_negative_drops')
}
\seealso{
\code{\link[ddpcrS3]{PPNP_ASSAy}}
}
\keyword{internal}

