% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{move_back}
\alias{move_back}
\title{Move columns to the back of a data.frame}
\usage{
move_back(df, cols)
}
\arguments{
\item{df}{A data.frame.}

\item{cols}{A vector of column names to move to the back}
}
\description{
This function is taken from daattali/rsalad R package.
}
\examples{
df <- data.frame(a = character(0), b = character(0), c = character(0))
move_back(df, "b")
move_back(df, c("b", "a"))
}
\keyword{internal}

