% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{move_front}
\alias{move_front}
\title{Move columns to the front of a data.frame (taken from daattali/rsalad)}
\usage{
move_front(df, cols)
}
\description{
Move columns to the front of a data.frame (taken from daattali/rsalad)
}
\keyword{internal}

