% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plate-attribs.R
\name{name}
\alias{name}
\alias{name<-}
\title{Plate name}
\usage{
name(plate)

name(plate) <- value
}
\arguments{
\item{plate}{A ddpcrPlate}

\item{value}{New name}
}
\value{
Plate name
}
\description{
Get or set the name of a dataset.
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir)
name(plate)
name(plate) <- "foo"
name(plate)
}
}

