% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-hex_positive_pnpp.R
\name{parent_plate_type.hex_positive_pnpp}
\alias{parent_plate_type.hex_positive_pnpp}
\title{Parent plate type of HEX-positive PNPP}
\usage{
\method{parent_plate_type}{hex_positive_pnpp}(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\description{
Parent plate type of HEX-positive PNPP
}
\keyword{internal}
