% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa.R
\docType{data}
\name{plate_types}
\alias{plate_types}
\title{Supported plate types}
\description{
Each ddPCR plate has a plate type which determines what type of analysis to run
on the data. \code{plate_types} is a list containing the plate types that are
supported. If no plate type is specified, the default type is
\code{ddpcr_plate}.\cr\cr
For full details on the differences between plate types or to learn how to
add a new plate type, \href{https://github.com/daattali/ddpcr}{see the package README}.
}
\seealso{
\code{\link[ddpcr]{new_plate}}
\code{\link[ddpcr]{fam_positive_pnpp}}
\code{\link[ddpcr]{hex_positive_pnpp}}
\code{\link[ddpcr]{custom_thresholds}}
\code{\link[ddpcr]{pnpp_experiment}}
\code{\link[ddpcr]{wildtype_mutant_pnpp}}
\code{\link[ddpcr]{ddpcr_plate}}
\code{\link[ddpcr]{type}}
}
\keyword{datasets}

