% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{plus_minus}
\alias{plus_minus}
\title{Get the two values that are equidistant from a specific number}
\usage{
plus_minus(x, y)
}
\arguments{
\item{x}{The middle number.}

\item{y}{The amount to deviate from the middle.}
}
\description{
Get the two values that are equidistant from a specific number
}
\keyword{internal}

