% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{pnpp_experiment}
\alias{pnpp_experiment}
\title{Plate type: PNPP experiment}
\description{
PNPP stands for "Positive-Negative;Negative-Positive", which is a reflection
of the clusters of non-empty droplets in the wells. Use this plate type when
your ddPCR data has three main clusters: double-negative (FAM-HEX-; empty droplets),
double-positive (FAM+HEX+; represent the "PP" in PNPP), and singly-positive
(either FAM+HEX- or HEX+FAM-; represent the "NP" in PNPP).
}
\details{
Every \code{pnpp_experiment} plate must define which dimension is its \emph{positive
dimension}.  The positive dimension is defined as the dimension that corresponds
to the dye that has a high fluoresence intensity in all non-empty droplets. The other
dimension is defined as the \emph{variable dimension}. For example, assuming
the HEX dye is plotted along the X axis and the FAM dye is along the Y axis,
a FAM+/FAM+HEX+ plate will have "Y" as its positive dimension because both
non-empty clusters have FAM+ droplets. Similarly, a HEX+/FAM+HEX+ plate will
have "X" as its positive dimension.

The positive dimension must be set in order to use a \code{pnpp_experiment}.
It is not recommended to use this type directly; instead you should use one
of the subtypes. If you do use this type directly, you must set the positive
dimension with \code{\link[ddpcr]{positive_dim}}.

\href{https://github.com/daattali/ddpcr#extend}{See the README} for
more information on plate types.
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir, type = plate_types$pnpp_experiment)
type(plate)
}
}
\seealso{
\code{\link[ddpcr]{plate_types}}
\code{\link[ddpcr]{fam_positive_pnpp}}
\code{\link[ddpcr]{hex_positive_pnpp}}
\code{\link[ddpcr]{wildtype_mutant_pnpp}}
\code{\link[ddpcr]{positive_dim}}
\code{\link[ddpcr]{wells_positive}}
\code{\link[ddpcr]{wells_negative}}
\code{\link[ddpcr]{analyze}}
\code{\link[ddpcr]{remove_failures}}
\code{\link[ddpcr]{remove_outliers}}
\code{\link[ddpcr]{remove_empty}}
\code{\link[ddpcr]{classify_droplets}}
\code{\link[ddpcr]{reclassify_droplets}}
}

