% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{positive_dim}
\alias{positive_dim}
\alias{positive_dim<-}
\title{Positive dimension in a PNPP experiment}
\usage{
positive_dim(plate)

positive_dim(plate) <- value
}
\arguments{
\item{plate}{A ddPCR plate.}

\item{value}{The dimension to set as the positive dimension ("X" or "Y")}
}
\description{
Get or set the dimension (X or Y) that has a high intensity in all non-empty
drops in a \code{PNPP_EXPERIMENT}.
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir = dir, type = plate_types$pnpp_experiment)
positive_dim(plate) <- "Y"
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}},
\code{\link[ddpcr]{variable_dim}}
}
\keyword{internal}

