% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{positive_dim_var}
\alias{positive_dim_var}
\title{Name of dye in positive dimension in PNPP experiment}
\usage{
positive_dim_var(plate)
}
\description{
Get the name of the dye that is along the positive dimension.
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}}\cr
\code{\link[ddpcr]{positive_dim}}
}
\keyword{internal}

