% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\name{positive_dim_var}
\alias{positive_dim_var}
\title{Name of variable of positive dimension in PPNP assay}
\usage{
positive_dim_var(plate)
}
\description{
Get the name of the variable that is along the dimension where all filled
drops should be positive.
}
\seealso{
\code{\link[ddpcr]{PPNP_ASSAy}},
\code{\link[ddpcr]{positive_dim}}
}
\keyword{internal}

