% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{remove_empty.ddpcr_plate}
\alias{remove_empty.ddpcr_plate}
\title{Analysis step: Remove empty droplets}
\usage{
\method{remove_empty}{ddpcr_plate}(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\description{
Analysis step: Remove empty droplets
}
\keyword{internal}

