% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pnpp_experiment-empty.R
\name{remove_empty.pnpp_experiment}
\alias{remove_empty.pnpp_experiment}
\title{Analysis step: Remove empty droplets}
\usage{
\method{remove_empty}{pnpp_experiment}(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\description{
Analysis step: Remove empty droplets
}
\keyword{internal}
