% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/failures.R
\name{remove_failures.ddpcr_plate}
\alias{remove_failures.ddpcr_plate}
\title{Removing failed wells}
\usage{
\method{remove_failures}{ddpcr_plate}(plate)
}
\description{
The algorithm for removing failed wells from a plate
}
\keyword{internal}

