% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/outliers.R
\name{remove_outliers.ddpcr_plate}
\alias{remove_outliers.ddpcr_plate}
\title{Remove outlier droplets}
\usage{
\method{remove_outliers}{ddpcr_plate}(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\description{
Remove outlier droplets
}
\keyword{internal}

