% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\name{variable_dim}
\alias{variable_dim}
\alias{variable_dim<-}
\title{Variable dimension in a PPNP assay}
\usage{
variable_dim(plate)

variable_dim(plate) <- value
}
\description{
Get or set the dimension (X or Y) that can have both high and low intensities
in the non-empty drops in a \code{PPNP_ASSAY}.
}
\examples{
dir <- system.file("sample_data", "small", package = "ddpcrS3")
plate <- new_plate(dir = dir, type = PPNP_ASSAY)
variable_dim(plate) <- "Y"
variable_dim(plate)
positive_dim(plate)
}
\seealso{
\code{\link[ddpcrS3]{PPNP_ASSAy}},
\code{\link[ddpcrS3]{positive_dim}}
}
\keyword{internal}

