% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{variable_dim_var}
\alias{variable_dim_var}
\title{Name of variable of variable dimension in PNPP experiment}
\usage{
variable_dim_var(plate)
}
\description{
Get the name of the variable that is along the dimension where the droplets
will cluster into two groups.
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}},
\code{\link[ddpcr]{variable_dim}}
}
\keyword{internal}

