% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\name{variable_dim_var}
\alias{variable_dim_var}
\title{Name of variable of variable dimension in PPNP assay}
\usage{
variable_dim_var(plate)
}
\description{
Get the name of the variable that is along the dimension where the droplets
will cluster into two groups.
}
\seealso{
\code{\link[ddpcrS3]{PPNP_ASSAy}},
\code{\link[ddpcrS3]{variable_dim}}
}
\keyword{internal}

