% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{variable_dim_var}
\alias{variable_dim_var}
\title{Name of dye in variable dimension in PNPP experiment}
\usage{
variable_dim_var(plate)
}
\description{
Get the name of the dye that is along the variable dimension.
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}}\cr
\code{\link[ddpcr]{variable_dim}}
}
\keyword{internal}

