% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{data}
\name{WELL_ID_REGEX}
\alias{WELL_ID_REGEX}
\title{regex for a well ID}
\format{An object of class \code{character} of length 1.}
\usage{
WELL_ID_REGEX
}
\description{
regex for a well ID
}
\keyword{internal}
