% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/type-ppnp_assay.R
\name{wells_positive}
\alias{wells_positive}
\title{Get wells containing mostly positive droplets in a PPNP assay}
\usage{
wells_positive(x)
}
\arguments{
\item{x}{A ddPCR plate of type PPNP_ASSAY}
}
\value{
Vector of wells with mostly positive drops.
}
\description{
In a PPNP assay, the two main non-empty clusters of drops are the negative
and positive clusters. The positive cluster appears in every well, while the
negative cluster is not necessarily in every well. \code{wells_positive}
returns a list of wells that have mostly positive drops.
}
\examples{
file <- system.file("sample_data", "small", "analyzed_ppnp.rds", package = "ddpcrS3")
plate <- load_plate(file)
plate \%>\% wells_positive
}
\seealso{
\code{\link[ddpcrS3]{PPNP_ASSAy}},
\code{\link[ddpcrS3]{wells_negative}}
}

