% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-pnpp_experiment.R
\name{wells_positive}
\alias{wells_positive}
\title{Get positive wells}
\usage{
wells_positive(plate)
}
\arguments{
\item{plate}{A ddPCR plate.}
}
\value{
Character vector with well IDs of positive wells
}
\description{
After a ddPCR plate of type \code{pnpp_experiment} has been analyzed,
get the wells that were deemed as having mostly positive droplets.
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir, type = plate_types$pnpp_experiment) \%>\% analyze
wells_positive(plate)
}
}
\seealso{
\code{\link[ddpcr]{pnpp_experiment}}
\code{\link[ddpcr]{wells_negative}}
}

