% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate.R
\name{wells_used}
\alias{wells_used}
\title{Get wells used in a ddPCR plate}
\usage{
wells_used(plate)
}
\arguments{
\item{plate}{A ddPCR plate}
}
\value{
List of wells that have any data in the given plate.
}
\description{
Get a list of the wells that have any data in a ddPCR plate.
}
\examples{
\dontrun{
plate <- new_plate(sample_data_dir(), type = plate_types$custom_thresholds)
wells_used(plate)
plate <- subset(plate, "A01:C05")
wells_used(plate)
} 
}
\seealso{
\code{\link[ddpcr]{subset.ddpcr_plate}}
}
