% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/type-wildtype_mutant_pnpp.R
\name{wildtype_mutant_pnpp}
\alias{wildtype_mutant_pnpp}
\title{Plate type: wildtype/mutant PNPP}
\description{
A plate of type \code{wildtype_mutant_pnpp} is a subtype of
\code{\link[ddpcr]{pnpp_experiment}} that assumes the double-positive cluster
denotes wildtype and the other non-empty cluster denotes mutant droplets.
There are two plate types that are subtypes of \code{wildtype_mutant_pnpp}:
\code{\link[ddpcr]{fam_positive_pnpp}} and \code{\link[ddpcr]{hex_positive_pnpp}}.
It is not recommended to use this type directly; instead you should use one
of the subtypes.
}
\details{
Plates with this type inherit the analysis steps from \code{pnpp_experiment}
and also have the following: \code{CLASSIFY}, \code{RECLASSIFY}

Plates with this type inherit the droplet clusters from \code{pnpp_experiment}
and also have the following: \code{UNDEFINED},
\code{FAILED}, \code{OUTLIER}, \code{RAIN}, \code{POSITIVE}, \code{NEGATIVE}.

\href{https://github.com/daattali/ddpcr#extend}{See the README} for
more information on plate types.
}
\examples{
\dontrun{
dir <- system.file("sample_data", "small", package = "ddpcr")
plate <- new_plate(dir, type = plate_types$wildtype_mutant_pnpp)
type(plate)
}
}
\seealso{
\code{\link[ddpcr]{plate_types}}
\code{\link[ddpcr]{fam_positive_pnpp}}
\code{\link[ddpcr]{hex_positive_pnpp}}
\code{\link[ddpcr]{pnpp_experiment}}
\code{\link[ddpcr]{analyze}}
\code{\link[ddpcr]{remove_failures}}
\code{\link[ddpcr]{remove_outliers}}
\code{\link[ddpcr]{remove_empty}}
\code{\link[ddpcr]{classify_droplets}}
\code{\link[ddpcr]{reclassify_droplets}}
}

